feat: add unit-tests for `merge_pressures`
Since the unit-testing is done using GTest, and hence C++ based, and `merge_pressures` is a "private" function inside `dive.c`, we use a C proxy function to expose it to the unit tests.

More generally, this can be used to test the rest of the functionality inside `dive.c` and the same approach can be used for the other C-based implementation files in Subsurface.

Signed-off-by: Michael Werle <micha@michaelwerle.com>